+++ asp.net-group/SmartLMS.Community/Program.cs
@@ -27,6 +27,9 @@ builder.Services.AddScoped<SmartLMS.Models.Security.IEncryptionService, SmartLMS
 builder.Services.AddScoped<ICommunityService, CommunityService>();
 builder.Services.AddScoped<IForumService, ForumService>();
 builder.Services.AddScoped<IBacklinkService, BacklinkService>();
+builder.Services.AddScoped<IResourceDiscussionService, ResourceDiscussionService>();
+builder.Services.AddScoped<IResourceRagService, ResourceRagService>();
+builder.Services.AddSingleton<IModerationService, ModerationService>();
 builder.Services.AddScoped<INotificationService, SmartLMS.Community.Services.CommunityNotificationService>();
 
 // 3. Redis Cache